Basic skeleton of the app
Added the basic skeleton of the app with Shadow.cljs.

Also added a very basic input handling from the user via the numeric
keyboard and displaying it back on the screen.

The idea is to then use this user's numeric input and apply the
algorithm to it to convert it to the desired mnemonic.