Added the main view with a picker
The reason scroll view was added is because the user should be able to
dismiss the keyboard and having a scroll view with ':scroll-enabled
false' allows the user to do that when they tap on the view.

I'm storing the input in a ratom, but maybe I can just use re-frame
for that instead.

The picker is populated with picker-items based on the input value the
user provides.