Display mnemonics
This enables displaying of mnemonics for the number provided, in
pickers.

It's not very pretty, but the functionality is there.

An obvious omission is the fact that if the number is too large, the
pickers do not all fit onto the screen.

I'm just committing this so that I can try some other UI elements,
that might take less screen space.

There is some ugly nested for-looping, with forced realisations with
doall. Some of it is due to the warning popups react-native was
throwing, but the most annoying one is the argument passed to
picker-options, which seems redundant to loop twice, but unless you do
it, it'll refuse to iterate over the mnemonics.