Added react-native-dropdown-menu
Added react-native-dropdown-menu. Overall it works, but there are
subtle issues.

For example when you stack multiple dropdowns and you click on the
very first dropdown, the options for it are shown below the last
dropdown.

To see what I mean, input the value 33780 and you'll get three
dropdown to play with.