Different ways to display dropdowns
Use native picker on Android and use react-native-picker-select for
iOS.

The reason for this "split" is because the native picker on iOS takes
too much screen real estate, whereas on Android it doesn't. So this
approach with react-native-picker-select is a workaround to allow the
user to be able to pick the mnemonics for numbers in a legible way.