Removed react-native-dropdown-menu
Tried it, didn't serve the purpose.

When you had a list of stacked dropdowns and you selected the items
from the first dropdown, it would display the items of the first
dropdown under the last dropdown. It looked very odd.

Also, found alternative ways to handle the dropdowns in both iOS &
Android.