Fixed the picker on Android
The issue was that I was using a ratom which did not keep state of
what was being picked from the drop down/picker.

Since I switched to using re-frame for storing the picked values, the
picker retains the value the user's choices.

The bug where if you have the input value of 3333, or any other number
where the digit "chunk" is duplicate, say 337833, you'd get the
chunks: [33, 78] (note the omitted trailing '33').