Added modal to be loaded on the initial load
Load the welcome modal explaining the game.

The modal is to be dismissed with a swipe in any direction.

Chose react-native-modal package as it gave you "more" out of the box:

- you can dismiss it with various swipe actions
- it automatically brings the modal to the front and puts the
  background in "shade"

These features can be developed using the stock modal & views, but
this is just easier to get going and it does what is needed.