Added a switch toggling & About screen invocation
Added a switch that can be toggled that doesn't do anything useful
yet, but it's there to show the possible functionality.

The other thing that was added to the Settings screen is the ability
to invoke the Welcome message, which will only be shown to the users
when the load the app initially.

I'll get back to making this a bit pretty, but for now it's just there
to show you where I'm at.