Prevent any undesired input from being used
Allow the user to put in any characters, but only the numerical ones
will be passed through and used.

The idea would be to maybe colour the input field and "shake" it to
let the user know that the value they've just added is not
permitted. But for now, this is a quick solution as per out chat.